feat: Saving assets data + updated returns endpoint.

+++ TradingBot/Dto/ReturnResponse.cs
@@ -5,7 +5,7 @@ namespace TradingBot.Dto;
 public sealed class ReturnResponse
 {
     [Required]
-    public required double Return { get; init; }
+    public required decimal Return { get; init; }
 
     [Required]
     public required DateTimeOffset Time { get; init; }
